src/test/java/ghstats/api/achievements/impl/DealWithItTest.java
package ghstats.api.achievements.impl;

import ghstats.api.achievements.api.AchievementUnlocked;
import ghstats.api.achievements.api.UnlockableAchievement;
import ghstats.api.integrations.github.api.GitCommit;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

import java.util.List;
import java.util.Optional;

class DealWithItTest extends BaseAchievementTest {
    @Test
    void shouldUnlock() {
        //given
        UnlockableAchievement achievement = new DealWithIt();
        List<GitCommit> commits = List.of(
                commitWithRefAndForced("not-on-master", "refs/heads/other", true),
                commitWithRefAndForced("on-master", "refs/heads/master", true),
        );

        //when
        Optional<AchievementUnlocked> check = achievement.unlock(commits);

        //expect
        Assertions.assertTrue(check.isPresent());
        Assertions.assertEquals("on-master", check.get().commit().id().value());
    }
}
